Create README.md
Vehicle Inventory System is a simple system that can add, remove, and edit vehicle items. Also, the system is useful as a model of how to construct a simple inventory system within Python. This simple project has a text file to store the current inventory details. Also, this is a simple cmd based project which is easy to understand and use.